feat: enhance and refactor Streamlit dashboard with improved UI and metrics
- Moves dashboard and its dependencies to `reporting/` for better organization.
- Integrates custom styling with `styles.css` and updates Streamlit theme colors.
- Replaces basic Streamlit charts with Plotly for richer, interactive visualizations.
- Adds 'total emissions' metric and new metric card components for key insights.
- Improves table rendering for better data presentation.
- Updates `.gitignore` with common Python environment entries and adjusts documentation paths.
