+++ reporting/requirements-dashboard.txt
@@ -1,3 +1,4 @@
 duckdb
 pandas
+plotly
 streamlit>=1.36
